pnpm

+++ bin/build.coffee
@@ -12,10 +12,11 @@ args = karg """
 
 build
     compile  . ? compile sources        . = true
-    npminst  . ? run npm install        . = true
+    npminst  . ? run npm install        . = false
+    pnpm     . ? run pnpm install       . = true
     builder  . ? electron-builder       . = true
     install  . ? move to /Applications  . = true
-    prune    . ? prune package          . = true . -P
+    prune    . ? prune package          . = false . -P
     start    . ? run executable         . = true
     verbose                             . = true
 """
@@ -56,14 +57,19 @@ try
     
     if args.compile then exec 'compile' 'node --trace-warnings ' + slash.join __dirname, 'konrad'
     if args.npminst then exec 'npminst' 'npm install'
+    if args.pnpm    then exec 'pnpm -i' 'pnpm install'
     if args.builder then exec 'builder' "#{slash.resolve('./node_modules/.bin/electron-builder')} --dir"
-    if args.install
+    if args.install and not slash.win()
         log kolor.y5 'install'
         appDir = "/Applications/#{slash.file exepth}"
         if slash.dirExists appDir
-            # klog "remove #{appDir}"
             fs.removeSync appDir
         fs.moveSync exepth, appDir
+        fs.removeSync slash.resolve 'dist'
+        exepth = appDir
+        fs.removeSync slash.join exepth, 'Contents/Resources/app/node_modules'
+        process.chdir slash.join exepth, 'Contents/Resources/app/'
+        childp.execSync 'pnpm i'
     if args.prune
         if args.verbose then log kolor.y4('prune')
         for d in ['inno' 'x64']
@@ -71,7 +77,7 @@ try
             if slash.dirExists dir
                 if args.verbose then log kolor.r5 dir
                 fs.removeSync dir
-                
+                 
         if prune = config.obj(pkgpth)?.build?.prune
             for d in prune        
                 dir = slash.join bindir, 'resources' 'app' d # needs to change on mac
@@ -80,9 +86,7 @@ try
                     fs.removeSync dir
                 else
                     log 'no path to prune' dir
-                
     if args.start
-        if args.install then exepth = appDir
         if args.verbose then log kolor.y3('start     '), kolor.w2 slash.tilde exepth
         if os.platform() == 'win32'
             childp.spawn exepth, encoding:'utf8' detached:true
